mockxchange e.g refactors mocking exchange
