+++ src/schema/__tests__/ecommerce/finalize_order_mutation.test.js
@@ -1,13 +1,6 @@
 /* eslint-disable promise/always-return */
 import { runQuery } from "test/utils"
-import {
-  makeExecutableSchema,
-  transformSchema,
-  RenameTypes,
-  RenameRootFields,
-} from "graphql-tools"
-import fs from "fs"
-import path from "path"
+import { mockxchange } from "test/fixtures/exchange/mockxchange"
 import sampleOrder from "test/fixtures/results/sample_order"
 import exchangeOrderJSON from "test/fixtures/exchange/order.json"
 
@@ -15,11 +8,6 @@ let rootValue
 
 describe("Finalize Order Mutation", () => {
   beforeEach(() => {
-    const typeDefs = fs.readFileSync(
-      path.resolve(__dirname, "../../../data/exchange.graphql"),
-      "utf8"
-    )
-
     const resolvers = {
       Mutation: {
         finalizeOrder: () => ({
@@ -29,51 +17,7 @@ describe("Finalize Order Mutation", () => {
       },
     }
 
-    const schema = makeExecutableSchema({
-      typeDefs,
-      resolvers,
-    })
-
-    // namespace schema similar to src/lib/stitching/exchange/schema.ts
-    const exchangeSchema = transformSchema(schema, [
-      new RenameTypes(name => {
-        return `Ecommerce${name}`
-      }),
-      new RenameRootFields((_operation, name) => `ecommerce_${name}`),
-    ])
-
-    const partnerLoader = sinon.stub().returns(
-      Promise.resolve({
-        id: "111",
-        name: "Subscription Partner",
-      })
-    )
-
-    const userByIDLoader = sinon.stub().returns(
-      Promise.resolve({
-        id: "111",
-        email: "bob@ross.com",
-      })
-    )
-
-    const authenticatedArtworkLoader = sinon.stub().returns(
-      Promise.resolve({
-        id: "hubert-farnsworth-smell-o-scope",
-        title: "Smell-O-Scope",
-        display: "Smell-O-Scope (2017)",
-        inventory_id: "inventory note",
-      })
-    )
-
-    const accessToken = "open-sesame"
-
-    rootValue = {
-      exchangeSchema,
-      partnerLoader,
-      userByIDLoader,
-      authenticatedArtworkLoader,
-      accessToken,
-    }
+    rootValue = mockxchange(resolvers)
   })
   it("fetches order by id", () => {
     const mutation = `
